+++ assist/fonts.js
@@ -1,18 +1,21 @@
-import { seek } from "lethil";
+import { seek, db, digit } from "lethil";
 import { resolve } from "path";
-// import util from "util";
+import util from "util";
+import * as fs from "fs";
 import ttfMeta from "ttfmeta";
 
 import config from "./anchor/env.js";
 
-// const { media } = config;
+/**
+ * @typedef {{file:string,name:string,version:string,family:string, view:number, download:number, restrict?:boolean, total?:string}} TypeFont
+ * property {{[k: string]:any}}
+ * @type {{[k: string]:TypeFont[]}}
+ */
+const fontData = {};
 
 export default class fonts {
-	/**
-	 * @property {{[k: string]:any}}
-	 * @type {{[k: string]:{[k: string]:any}}}
-	 */
-	store = {};
+	catalogue = ["primary", "secondary", "external"];
+
 	/**
 	 * @param {string} type
 	 */
@@ -21,7 +24,9 @@ export default class fonts {
 	}
 
 	/**
+	 * font root directory
 	 * @param {...string} fileName
+	 * @return {string}
 	 */
 	root(...fileName) {
 		return resolve(config.media, "fonts", fileName.join("/"));
@@ -30,94 +35,129 @@ export default class fonts {
 	/**
 	 * @param {string} fileName
 	 */
-	fileJSON(fileName = this.type) {
+	fileJSON(fileName) {
 		return this.root(fileName + ".json");
 	}
 
 	/**
-	 * @param {string} fileName
+	 * Read JSON by catalogue
+	 * @param {string} catalogue
+	 * @returns {Promise<TypeFont[]>}
 	 */
-	async read(fileName = this.type) {
-		if (this.store.hasOwnProperty(fileName)) {
-			return this.store[fileName];
+	async read(catalogue) {
+		if (fontData.hasOwnProperty(catalogue)) {
+			return fontData[catalogue];
 		} else {
-			let file = this.fileJSON(fileName);
-			this.store[fileName] = [];
+			// let types = this.catalogue.indexOf(this.type);
+			// if (types < 0) {
+			// 	return "catalogue must be provided";
+			// }
 
+			let file = this.fileJSON(catalogue);
+			fontData[catalogue] = [];
 			return seek
 				.read(file)
-				.then(e => (this.store[fileName] = JSON.parse(e.toString())))
+				.then(e => (fontData[catalogue] = JSON.parse(e.toString())))
 				.catch(() => {});
 		}
 	}
 
 	/**
-	 * @param {string} fileName
+	 * @param {string} catalogue
 	 */
-	write(fileName) {
-		if (this.store.hasOwnProperty(fileName)) {
-			let file = this.fileJSON(fileName);
-			seek
-				.write(file, JSON.stringify(this.store[fileName]))
+	async write(catalogue) {
+		if (fontData.hasOwnProperty(catalogue)) {
+			let file = this.fileJSON(catalogue);
+			return seek
+				.write(file, JSON.stringify(fontData[catalogue]))
 				.then(() => true)
 				.catch(() => false);
 		}
-	}
-
-	/**
-	 * @param {string} name
-	 */
-	fileFont(name) {
-		return this.root(this.type, name);
+		return false;
 	}
 
 	/**
 	 * @param {string} fileName
 	 */
-	async view(fileName) {
-		var response = {};
-		var item = null;
-		if (this.type && fileName) {
-			var file = this.fileFont(fileName);
-			await this.read(this.type);
-
-			if (this.store.hasOwnProperty(this.type)) {
-				var index = this.store[this.type].findIndex(x => x.file == fileName);
-				if (index >= 0) {
-					item = this.store[this.type][index];
-					if (item instanceof Object) {
-						this.store[this.type][index].view++;
-						this.sortable();
-						this.write(this.type);
-					}
-				}
-			}
-			await ttfMeta.promise(file).then(async e => {
-				if (e && e.hasOwnProperty("tables")) {
-					response = await this.view_response(e.tables.name);
-					if (item instanceof Object && !item.hasOwnProperty("restrict")) {
-						response.unrestrict = true;
-					}
-				}
-			});
-		}
-		return response;
+	fileFont(fileName) {
+		return this.root(this.type, fileName);
 	}
 
 	/**
+	 * View meta
+	 * @param {string} fileName
+	 */
+	async view(fileName) {
+		/**
+		 * @type {{[k:string]:any}}
+		 */
+		var info = {};
+
+		for (let index = 0; index < this.catalogue.length; index++) {
+			const cate = this.catalogue[index];
+			if (!fontData.hasOwnProperty(cate)) {
+				await this.read(cate);
+			}
+		}
+
+		/**
+		 * @type {{ hasOwnProperty: (arg0: string) => any; } | null}
+		 */
+		var item = null;
+		if (this.type && fileName) {
+			var file = this.fileFont(fileName);
+
+			if (fontData.hasOwnProperty(this.type)) {
+				var index = fontData[this.type].findIndex(e => e.file == fileName);
+
+				if (index >= 0) {
+					item = fontData[this.type][index];
+					if (item instanceof Object) {
+						fontData[this.type][index].view++;
+						this._sortBySum(this.type);
+
+						db.mysql.query(
+							"INSERT INTO ?? SET file=?, view=1 ON DUPLICATE KEY UPDATE view = view + 1;",
+							[config.table.fonts, fileName]
+						);
+					}
+				}
+			}
+			await ttfMeta
+				.promise(file)
+				.then(async e => {
+					if (e && e.hasOwnProperty("tables")) {
+						info = await this._metaDetail(e.tables.name);
+						if (item instanceof Object && !item.hasOwnProperty("restrict")) {
+							info.unrestrict = true;
+						}
+					}
+				})
+				.catch(e => {
+					info = {};
+				});
+		}
+		return Object.assign(info, fontData);
+	}
+
+	/**
+	 * Download data
 	 * @param {string} fileName
 	 */
 	async download(fileName) {
 		if (this.type && fileName) {
 			await this.read(this.type);
-			if (this.store.hasOwnProperty(this.type)) {
-				var index = this.store[this.type].findIndex(x => x.file == fileName);
+			if (fontData.hasOwnProperty(this.type)) {
+				var index = fontData[this.type].findIndex(e => e.file == fileName);
 				if (index >= 0) {
-					var item = this.store[this.type][index];
+					var item = fontData[this.type][index];
 					if (item instanceof Object && !item.hasOwnProperty("restrict")) {
-						this.store[this.type][index].download++;
-						this.sortable();
-						this.write(this.type);
+						fontData[this.type][index].download++;
+						this._sortBySum(this.type);
+						db.mysql.query(
+							"INSERT INTO ?? SET file=?, download=1 ON DUPLICATE KEY UPDATE download = download + 1;",
+							[config.table.fonts, fileName]
+						);
 						return this.fileFont(fileName);
 					}
 				}
@@ -129,8 +169,11 @@ export default class fonts {
 	/**
 	 * @param {{[k:string]:any}} e
 	 */
-	async view_response(e) {
-		var abc = {
+	async _metaDetail(e) {
+		/**
+		 * @type {any}
+		 */
+		const tpl = {
 			0: "Copyright",
 			1: "Font Family",
 			2: "Font Subfamily",
@@ -150,26 +193,31 @@ export default class fonts {
 			16: "Name"
 			// 17: ''
 		};
-		// info description license, url
-		var response = {
+
+		/**
+		 * info description license, url
+		 * @type {{[k: string]:any}}
+		 */
+		const response = {
 			// title:null,
 			// keywords:null,
 			// description:null,
 			info: []
-			// description:[],
-			// license:[],
+			// description: [],
+			// license: []
 			// url:[]
 		};
 
 		for (const key in e) {
 			if (e.hasOwnProperty(key)) {
 				const i = parseInt(key);
-				const context = e[i].trim();
+				// itemFamily.replace(/\u0000/g, ""),
+				const context = e[i].replace(/\u0000/g, "").trim();
 				var paragraphs = context
 					.replace("~\r\n?~", "\n")
 					.split("\n")
-					.map(i => i.trim())
-					.filter(i => i != null && i != "");
+					.map((/** @type {string} */ i) => i.trim())
+					.filter((/** @type {string | null} */ i) => i != null && i != "");
 				if (paragraphs.length > 1) {
 					var id = i == 10 ? "definition" : "license";
 					response[id] = [];
@@ -182,7 +230,10 @@ export default class fonts {
 									? "p"
 									: "h3"
 								: "p";
-							response[id].push({ tag: tagName, text: text });
+							response[id].push({
+								tag: tagName,
+								text: text
+							});
 						}
 					}
 				} else if (context) {
@@ -190,11 +241,14 @@ export default class fonts {
 						/^s?https?:\/\/[-_.!~*'()a-zA-Z0-9;\/?:\@&=+\$,%#]+$/.test(context)
 					) {
 						if (!response.hasOwnProperty("url")) response.url = [];
-						response.url.push({ href: context, text: context });
+						response.url.push({
+							href: context,
+							text: context
+						});
 						// response.url = [...new Set(response.url)];
 						// response.url = response.url.filter((v, i, a) => a.indexOf(v) === i);
 					} else if (i > 0 && i < 6) {
-						var className = abc[i].replace(" ", "-").toLowerCase();
+						var className = tpl[i].replace(" ", "-").toLowerCase();
 						var tagName = "h" + i;
 						response.info.push({
 							tag: tagName,
@@ -202,15 +256,17 @@ export default class fonts {
 							text: context
 						});
 					} else {
-						if (abc.hasOwnProperty(i)) {
+						if (tpl.hasOwnProperty(i)) {
 							if (i == 0 || i == 7) {
 								if (!response.hasOwnProperty("definition"))
 									response.definition = [];
 								var paragraphs = context
 									.replace(/---+/, "\n")
 									.split("\n")
-									.map(i => i.trim())
-									.filter(i => i != null && i != "");
+									.map((/** @type {string} */ i) => i.trim())
+									.filter(
+										(/** @type {string | null} */ i) => i != null && i != ""
+									);
 								for (const eP in paragraphs) {
 									if (paragraphs.hasOwnProperty(eP)) {
 										var text = paragraphs[eP].trim();
@@ -219,7 +275,10 @@ export default class fonts {
 												? "p"
 												: "h3"
 											: "p";
-										response.definition.push({ tag: tagName, text: text });
+										response.definition.push({
+											tag: tagName,
+											text: text
+										});
 									}
 								}
 							} else if (i == 13) {
@@ -229,9 +288,12 @@ export default class fonts {
 										? "p"
 										: "h3"
 									: "p";
-								response.license.push({ tag: tagName, text: context });
+								response.license.push({
+									tag: tagName,
+									text: context
+								});
 							} else {
-								var className = abc[i].replace(" ", "-").toLowerCase();
+								var className = tpl[i].replace(" ", "-").toLowerCase();
 								response.info.push({
 									tag: "p",
 									class: className,
@@ -256,9 +318,135 @@ export default class fonts {
 		return response;
 	}
 
-	sortable() {
-		return (this.store[this.type] = this.store[this.type].sort(function(a, b) {
-			return b.download + b.view - (a.download + a.view);
+	/**
+	 * Sort by the sum of (view+download) and Update fontData
+	 * @param {string} cate
+	 */
+	_sortBySum(cate) {
+		return (fontData[cate] = fontData[cate].sort((a, b) => {
+			return b.view + b.download - (a.view + a.download);
 		}));
+	}
+
+	/**
+	 * Font rebuild catalogue JSON
+	 * param {string?} fontParam
+	 */
+	async scan() {
+		for (let index = 0; index < this.catalogue.length; index++) {
+			const cate = this.catalogue[index];
+			const status = await this.categorize(cate);
+			console.log("> font", cate, status);
+		}
+	}
+
+	/**
+	 * use in reading if no fontData
+	 * or scan
+	 * @param {string} cate - cat/type not need filter as it's internal
+	 */
+	async categorize(cate) {
+		/**
+		 * @type{TypeFont[]}
+		 */
+		fontData[cate] = [];
+		var directory = this.root(cate);
+		var read_dir = util.promisify(fs.readdir);
+		// this.read("restrict");
+
+		const files = await read_dir(directory);
+
+		for (let index = 0; index < files.length; index++) {
+			const fileName = files[index];
+
+			const o = await ttfMeta.promise(this.root(cate, fileName));
+			if (o && o.hasOwnProperty("meta")) {
+				const org = await db.mysql.query(
+					"SELECT view, download, restricted FROM ?? WHERE file=?;",
+					[config.table.fonts, fileName]
+				);
+				var countView = 0;
+				var countDownload = 0;
+				var isrestrict = false;
+
+				if (org.length) {
+					countView = org[0].view;
+					countDownload = org[0].download;
+
+					isrestrict = org[0].restricted > 0;
+				}
+
+				let oj = o.meta.property;
+				let itemFamily = oj.find(e => e.name == "font-family")?.text || "";
+				let itemVersion = oj.find(e => e.name == "version")?.text || "";
+				let itemSub = oj.find(e => e.name == "font-subfamily")?.text || "";
+				/**
+				 * @type {TypeFont}
+				 */
+				var item = {
+					file: fileName,
+					name: itemFamily.replace(/\u0000/g, ""),
+					version: itemVersion.replace(/\u0000/g, ""),
+					family: itemSub.replace(/\u0000/g, ""),
+					view: countView,
+					download: countDownload,
+					total: digit(countView + countDownload).shorten()
+				};
+
+				if (isrestrict) {
+					item.restrict = true;
+				}
+				fontData[cate].push(item);
+			}
+		}
+
+		this._sortBySum(cate);
+		return this.write(cate);
+	}
+
+	/**
+	 * Integration counter from flat to db
+	 * @returns {Promise<string>}
+	 */
+	async scan_integration() {
+		let types = this.catalogue.indexOf(this.type);
+		if (types < 0) {
+			return "catalogue must be provided";
+		}
+		var json = await this.read(this.type);
+		var directory = this.root(this.type);
+		var read_dir = util.promisify(fs.readdir);
+		this.read("restrict");
+
+		const files = await read_dir(directory);
+
+		for (let index = 0; index < files.length; index++) {
+			const fileName = files[index];
+			let org = json.find(e => e.file == fileName);
+
+			let countView = 0;
+			let countDownload = 0;
+			var isrestrict = 0;
+			if (org) {
+				countView = org.view;
+				countDownload = org.download;
+				if (org.hasOwnProperty("restrict") && org.restrict == true) {
+					isrestrict = 1;
+				}
+			}
+			db.mysql.query(
+				"INSERT INTO ?? SET file=?, types=?, view=?, download=?, restricted=? ON DUPLICATE KEY UPDATE view = view + 1;",
+				[
+					config.table.fonts,
+					fileName,
+					types,
+					countView,
+					countDownload,
+					isrestrict
+				]
+			);
+		}
+
+		return "done";
 	}
 }